Move js files around
